% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esp.R
\name{theme_esp}
\alias{theme_esp}
\title{Theme inspired by Economic Security Project graphics}
\usage{
theme_esp(
  base_size = 12,
  base_family = "Publico Banner",
  background = c("beige", "white")
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{background}{\code{"white"} (default) or \code{"beige"}.}
}
\description{
A minimal theme with ESP brand colours, grid, and typography.  It builds on
\strong{\link[ggplot2]{theme_minimal}}.  Use together with
\code{\link[=scale_colour_esp]{scale_colour_esp()}} / \code{\link[=scale_fill_esp]{scale_fill_esp()}}.
}
\concept{themes esp}
