% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esp.R
\docType{data}
\name{esp_navy}
\alias{esp_navy}
\title{ESP Primary Color (Navy)}
\format{
An object of class \code{character} of length 1.
}
\usage{
esp_navy
}
\description{
A standalone color value for quick use.
}
\keyword{datasets}
