% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esp.R
\name{esp_navy}
\alias{esp_navy}
\title{Warm Navy brand colour}
\usage{
esp_navy()
}
\value{
Character vector of length 1.
}
\description{
Returns the hex code \code{"#2c3254"}.
}
\seealso{
Other colour esp: 
\code{\link{esp_pal}()},
\code{\link{scale_colour_esp}()}
}
\concept{colour esp}
