% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esp.R
\docType{data}
\name{esp_pal}
\alias{esp_pal}
\title{ESP Color Palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
esp_pal
}
\description{
Named vector of ESP-branded colors.
}
\keyword{datasets}
