% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esp.R
\name{esp_pal}
\alias{esp_pal}
\title{ESP colour palette (discrete)}
\usage{
esp_pal()
}
\description{
Returns a manual palette function using the four ESP brand colours in order:
Warm Navy, Warm Red, Soft Green, Deep Purple.
}
\seealso{
Other colour esp: 
\code{\link{esp_navy}()},
\code{\link{scale_colour_esp}()}
}
\concept{colour esp}
